requirements.txt
# Web scraping
newsapi-python==0.2.6
praw==7.7.1
playwright==1.40.0
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2

# NLP and ML
bertopic==0.15.0
spacy==3.7.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
umap-learn==0.5.4
textblob==0.17.1
vaderSentiment==3.3.2
transformers==4.36.0

# Time series forecasting
prophet==1.1.4
statsmodels==0.14.0
tensorflow==2.15.0

# Data processing
pandas==2.1.4
numpy==1.24.3
scipy==1.11.4

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
# sqlite3

# Visualization and Dashboard
streamlit==1.28.2
plotly==5.17.0
dash==2.14.2
matplotlib==3.8.2
seaborn==0.13.0

# Scheduling and utilities
schedule==1.2.0
python-dotenv==1.0.0
loguru==0.7.2
tqdm==4.66.1